Move brent focuser back in to oughtafocus plugin
